CI : remove dead helpers and scope PF runtime
